removing the dust
Made some changes on the Game Engine logics, now every feature that was part from the gameEngine.py has become its own class/module on the project.

[+] gameEngine.py: Inventory.showInventory() => inventorySystem.py: Inventory.showInventory()

[+] gameEngine.py: Native.updateScreen() => mapHandler.py: MapHandler.UI().update(args)

[+] dataStock.py: DataStock.seeAllData() -> now every data from the data files (json) in the game save can be seen with one command (still in development)

[+] module.py -> new module system: now scripts can be converted into features of the game, makes easy for me to integrate functionality to the game, such as the map handler which creates a new screen based on a set of data, stores the screen state and updates this screen if needed.

[*] main.py => jarparur.py

[*] gameEngine.py: start() -> that's all now, no more polluted code!

That's it, a lot of work done here, hello again Jarparur!
