+++ jarparur.py
@@ -11,7 +11,7 @@ def mainStructure():
     print(art.gameTitle)
 
     while True:
-        action = input("\n| Novo Jogo     [1] |\n| Carregar Jogo [2] |\n| Sair          [3] | \n\n>>> ")
+        action = input("\n| Novo Jogo     [1] |\n| Carregar Jogo [2] |\n| Sair [3] | \n\n>>> ")
         actions = ['1', '2', '3']
 
         if action in actions:
