+++ saves/Einar Ragnarsson/Einar Ragnarsson.json
@@ -1,1 +1,44 @@
-{"playerName": "Einar", "playerRace": "nulr", "playerClass": "bjoretNen", "playerLanguage": "nulrfanghe", "playerAge": 25, "playerHeight": 1.82, "playerHealth": 60, "playerMana": 50, "playerArmor": 2, "playerLevel": 1, "playerXP": 0, "playerVelocity": 8, "playerStrength": 8, "playerDexterity": 12, "playerConstitution": 12, "playerIntelligence": 15, "playerWisdom": 15, "playerCharisma": 6, "playerStartPoint": "nulr", "playerFavoriteWeapon": "M\u00e1gico", "playerInventory": {"head": "", "torso": "5111", "arms": "", "hands": "", "legs": "", "foot": "501", "slots": ["6151", "", "", "", "", "", "", ""], "bag": ["", "", ""]}, "playerAbilities": {"ability1": "151", "ability2": "152"}}
+{
+    "playerName": "Einar", 
+    "playerRace": "nulr", 
+    "playerClass": "bjoretNen", 
+    "playerLanguage": "nulrfanghe", 
+    "playerAge": 25, 
+    "playerHeight": 1.82, 
+    "playerHealth": 60, 
+    "playerMana": 50, 
+    "playerArmor": 2, 
+    "playerLevel": 1, 
+    "playerXP": 0, 
+    "playerVelocity": 8, 
+    "playerStrength": 8, 
+    "playerDexterity": 12, 
+    "playerConstitution": 12, 
+    "playerIntelligence": 15, 
+    "playerWisdom": 15, 
+    "playerCharisma": 6, 
+    "playerStartPoint": "nulr", 
+    "playerFavoriteWeapon": "M\u00e1gico",
+
+    "playerInventory": {
+        "head": "", 
+        "torso": "5111", "arms": "", 
+        "hands": "", 
+        "legs": "", 
+        "foot": "501", 
+        "slots": ["6151", "", "", "", "", "", "", ""], 
+        "bag": ["", "", ""]
+    }, 
+
+    "playerAbilities": {
+        "ability1": "151", 
+        "ability2": "152"
+    },
+
+    "playerWorld": {
+        "worldRegion": "test",
+        "localRegion": "testArea",
+        "x": 0,
+        "y": 0
+    }    
+}
