+++ saves/magoteste/magoteste.json
@@ -1,0 +1,1 @@
+{"playerName": "Mago Teste", "playerRace": "human", "playerClass": "wizard", "playerLanguage": "nulrfanghe", "playerAge": 25, "playerHeight": 1.95, "playerHealth": 60, "playerMana": 50, "playerArmor": 2, "playerLevel": 1, "playerXP": 0, "playerVelocity": 8, "playerStrength": 6, "playerDexterity": 10, "playerConstitution": 6, "playerIntelligence": 15, "playerWisdom": 15, "playerCharisma": 12, "playerStartPoint": "nulr", "playerFavoriteWeapon": "Utiliz\u00e1vel", "playerInventory": {"head": "", "torso": "5111", "arms": "", "hands": "", "legs": "", "foot": "501", "slots": ["6111", "", "", "", "", "", "", ""], "bag": ["", "", ""]}, "playerAbilities": {"ability1": "111", "ability2": "112"}}
