+++ saves/yes/worldData.json
@@ -1,0 +1,1 @@
+{"worldName": "Jarparur", "worldAge": 20000, "worldPopulation": 100000, "worldType": "Planet", "nulr": {"name": "Nulr", "description": "Uma bela regi\u00e3o ao norte de Jarparunn, separada do resto do mundo pelo muro colossal de Grenk\u00ebparur.", "language": "nulrfanghe"}, "grenkeparur": {"name": "Grenkeparur", "description": "Um gigantesco muro de gelo, destino daqueles que foram contra as leis do Sul, uma gigantesca pris\u00e3o.", "language": "espocioComum"}, "desertoDosOssos": {"name": "Deserto dos Ossos", "description": "Quente, \u00e1rido e lar das criaturas mais ferozes de Jarparunn. Voc\u00ea vai precisar de muita \u00e1gua aqui.", "language": "espocioAntigo"}, "sul": {"name": "O Sul", "description": "Terra dos humanos e dos elfos da floresta, possui imensos reinos e pol\u00edticas, e claro, muita magia arcana.", "language": "espocioComum"}}
