+++ saves/yes/yes.json
@@ -1,0 +1,1 @@
+{"playerName": "yes", "playerRace": "human", "playerClass": "warrior", "playerLanguage": "espocioComum", "playerAge": 23, "playerHeight": 1.45, "playerHealth": 120, "playerMana": 10, "playerArmor": 2, "playerLevel": 1, "playerXP": 0, "playerVelocity": 5, "playerStrength": 15, "playerDexterity": 8, "playerConstitution": 14, "playerIntelligence": 6, "playerWisdom": 12, "playerCharisma": 12, "playerStartPoint": "nulr", "playerFavoriteWeapon": "Arma de uma m\u00e3o", "playerInventory": {"head": "", "torso": "", "arms": "", "hands": "", "legs": "402", "foot": "501", "slots": ["101", "", "", "", "", "", "", ""], "bag": ["", "", ""]}, "playerAbilities": {"ability1": "", "ability2": ""}}
